requirements.txt
pandas~=1.3.0
xmltodict~=0.12.0
zxing~=0.12
numpy~=1.21.0
opencv-python~=4.5.2.54
requests~=2.25.1
python-telegram-bot~=13.7